backend/requirements.txt
Flask==3.0.0
flask-cors==4.0.0
firebase-admin==6.3.0
psutil==5.9.6
python-dotenv==1.0.0